Moved files Remove unused dependencies
