+++ package.json
@@ -6,17 +6,12 @@
     "@material-ui/core": "^3.8.1",
     "@material-ui/icons": "^3.0.1",
     "axios": "^0.18.0",
-    "cross-env": "^5.2.0",
     "prop-types": "^15.6.2",
     "react": "^16.7.0",
-    "react-app-polyfill": "^0.2.0",
     "react-dom": "^16.7.0",
     "react-frontload": "^1.0.3",
-    "react-helmet": "^5.2.0",
     "react-loadable": "^5.5.0",
     "react-redux": "^6.0.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.2",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.7",
@@ -50,7 +45,6 @@
     "enzyme-adapter-react-16": "^1.7.1",
     "enzyme-to-json": "^3.3.5",
     "eslint-config-airbnb": "^17.1.0",
-    "eslint-import-resolver-babel-module": "^4.0.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.12.1",
