+++ src/app/components/__tests__/Layout.snapshot-spec.js
