+++ src/app/state/todos/middleware.js
@@ -1,4 +1,4 @@
-import fetcher from '../../utils/fetcher';
+import fetcher from '../../../utils/fetcher';
 import { loadTodosFail, loadTodosSuccess } from './actions';
 import * as actionTypes from './action-types';
